# [benchmark] Add some benchmarks for String breadcrumbs