--- a/benchmark/CMakeLists.txt
+++ b/benchmark/CMakeLists.txt
@@ -42,6 +42,7 @@ set(SWIFT_BENCH_MODULES
     single-source/BinaryFloatingPointConversionFromBinaryInteger
     single-source/BinaryFloatingPointProperties
     single-source/BitCount
+    single-source/Breadcrumbs
     single-source/ByteSwap
     single-source/COWTree
     single-source/COWArrayGuaranteedParameterOverhead
@@ -153,7 +154,6 @@ set(SWIFT_BENCH_MODULES
     single-source/StaticArray
     single-source/StrComplexWalk
     single-source/StrToInt
-    single-source/StringBreadcrumbs
     single-source/StringBuilder
     single-source/StringComparison
     single-source/StringEdits