… benchmark/single-source/StringBreadcrumbs.swift