This adds some benchmarks for `String` breadcrumbing, via low-level stdlib entry points exposed by `String`.

- `StringBreadcrumbs.UTF16ToIndex`: Convert some random UTF-16 offsets into String indices.
- `StringBreadcrumbs.IndexToUTF16`: Convert some random String indices into UTF-16 offsets.
- `StringBreadcrumbs.UTF16ToIndexRange`: Split a string into a bunch of random slices and convert their UTF-16 offsets into String index ranges.
- `StringBreadcrumbs.IndexToUTF16Range`: Do the same in reverse direction.
- `StringBreadcrumbs.CopyUTF16CodeUnits`: Copy UTF-16 code units from random ranges.
- `StringBreadcrumbs.MutatedUTF16ToIndex`: This is like `UTF16ToIndex` but appends to the string after every index conversion. In effect, this tests breadcrumb creation performance.
- `StringBreadcrumbs.MutatedIndexToUTF16`: This is like `UTF16ToIndex` but appends to the string after every index conversion. In effect, this tests breadcrumb creation performance.

https://bugs.swift.org/browse/SR-9226