@lorentey 2018-11-26
The CopyUTF16CodeUnits benchmark here will crash until https://github.com/apple/swift/pull/20768 lands.