@palimondo 2018-11-26
@lorentey #20667 has landed. You should be able to run these now with swift-ci, too.