@lorentey 2018-11-26
Please test with the following PR: https://github.com/apple/swift/pull/20768
@please benchmark