@lorentey 2018-11-26
https://github.com/apple/swift/pull/20768
@swift-ci benchmark