@lorentey 2018-11-26
@palimondo The doctor complained about the names and the "very wide range of memory used", but I scaled things so that it likes the timings.