@lorentey 2018-11-26
@milseman I did some spot checks -- the heavy stack frames were inside the target entry points and looked breadcrumby enough. I wasn't looking for missed optimization opportunities though. 

I'll have a more detailed look tomorrow, just to make sure these are measuring what I think they're measuring.