@palimondo 2018-11-26
> The doctor complained about the names and the "very wide range of memory used"
Interesting. Could you re-check those with `--verbose` flag, to see how bad it is?
Tip: You can also run just select benchmarks by their test numbers, which can come in handy if you want to focus on sub-ranges, like this:
````
$ ./Benchmark_Driver run -i 3 --output-dir log {204..217} {670..683}
````