@lorentey 2018-11-26
@palimondo Here's a quick sample of the memory-related output:

```
memory: 'StringBreadcrumbs.CopyUTF16CodeUnits.shortASCII' has very wide range of memory used between independent, repeated measurements.
memory: StringBreadcrumbs.CopyUTF16CodeUnits.shortASCII mem_pages [i1, i2]:  min=[42, 46] 𝚫=4 R=[4, 17]
memory: 'StringBreadcrumbs.IndexToUTF16.longASCII' has very wide range of memory used between independent, repeated measurements.
memory: StringBreadcrumbs.IndexToUTF16.longASCII mem_pages [i1, i2]:  min=[1108, 1119] 𝚫=11 R=[44, 31]
memory: 'StringBreadcrumbs.IndexToUTF16.longMixed' has very wide range of memory used between independent, repeated measurements.
memory: StringBreadcrumbs.IndexToUTF16.longMixed mem_pages [i1, i2]:  min=[323, 326] 𝚫=3 R=[36, 36]
memory: 'StringBreadcrumbs.IndexToUTF16Range.longASCII' has very wide range of memory used between independent, repeated measurements.
memory: StringBreadcrumbs.IndexToUTF16Range.longASCII mem_pages [i1, i2]:  min=[1116, 1125] 𝚫=9 R=[33, 29]
```