@milseman 2018-11-26
@swift-ci please benchmark