@palimondo 2018-11-27
@milseman 
> where is the 40-character limit coming from and is there any way to get more? Is it Github's in-PR renderer that's constraining this?

Yes, primarily GitHub's fixed table width. As the above report shows, these name at 48 chars are filling the available entire space. But this is a best case scenario, as the `-O` runtimes are under 1000 μs . When you combine this with longer runtimes and all the benchmarks, it triggers horizontal scrolling.

Also formatting the result table in console (`Benchmark_Driver`'s output as run through `build_script`). And using all of benchmark reports on mobile devices, see https://github.com/apple/swift/pull/20334#discussion_r234527138.