@palimondo 2018-11-27
The variance in `IndexToUTF16` group worries me:


TEST | MIN | MAX | MEAN | MAX_RSS
-- | -- | -- | -- | --
StringBreadcrumbs.IndexToUTF16.longASCII | 660 | 2211 | 1177 | —
StringBreadcrumbs.IndexToUTF16.longMixed | 791 | 1312 | 965 | —
StringBreadcrumbs.IndexToUTF16Range.longASCII | 306 | 1786 | 800 | —
StringBreadcrumbs.IndexToUTF16Range.longMixed | 387 | 390 | 388| —

Only the `longMixed` look fine. This might be background process interference, with `run_smoke_bench`. It still uses averaging… instead of going with `--num-iters=1`.

@lorentey Could you please check locally the variance isn't this crazy, when measured properly?
````
$ ./Benchmark_Driver run -i 3 -f StringBreadcrumbs.IndexToUTF16Range