@lorentey 2018-11-28
`IdxToUTF16` looks fine to me locally:

```
$ .../Benchmark_Driver run -i 3 -f Breadcrumbs.IdxToUTF16
#,TEST,SAMPLES,MIN(μs),Q1(μs),MEDIAN(μs),Q3(μs),MAX(μs),MAX_RSS(B)
45,Breadcrumbs.IdxToUTF16.longASCII,55,421,433,434,435,452,4575232
46,Breadcrumbs.IdxToUTF16.longMixed,58,555,572,574,577,600,1310720
47,Breadcrumbs.IdxToUTF16Range.longASCII,57,202,209,209,210,214,4575232
48,Breadcrumbs.IdxToUTF16Range.longMixed,58,290,299,299,301,311,1327104

Total performance tests executed: 4
```