@lorentey 2018-11-28
Rebased, applied review notes, cleaned up some things.

@swift-ci please smoke test