@lorentey 2018-11-28
@swift-ci please smoke test