@lorentey 2018-11-28
@milseman I profiled all tests looking for anomalies; they seem to work as intended. I did not notice anything left unspecialized, but it seems noteworthy that the `IdxToUTF16` and `IdxToUTF16Range` tests spent about half of their time in BidirectionalCollection's index navigation, and only a third in `getBreadcrumb(forIndex:)`. Some internal heuristic may not work right for these tests.