@lorentey 2018-11-28
Please test with the following PR:

https://github.com/apple/swift/pull/20807
@swift-ci please benchmark