@palimondo 2018-11-28
@lorentey BTW, brace for `BenchmarkDoctor` complaining about names not being in `UpperCamelCase`, since the new naming convention isn't in yet… could you maybe help me finish and merge that in the meantime?