@palimondo 2018-11-28
@lorentey It looks like it didn't run the benchmarks with the other PR. I guess this only works for `test`? // cc @eeckstein