@palimondo 2018-11-28
Benchmark | MIN | MAX | MEAN
-- | --: | --: | --: 
Breadcrumbs.IdxToUTF16.longASCII | 586 | 2077 | 1083
Breadcrumbs.IdxToUTF16.longMixed | 714 | 724 | 720
Breadcrumbs.IdxToUTF16Range.longASCII | 289 | 1756 | 778
Breadcrumbs.IdxToUTF16Range.longMixed | 371 | 373 | 372

🧐`longMixed` look great, but `longASCII` is weirding me out… You said it [looked fine locally](https://github.com/apple/swift/pull/20769#issuecomment-442569784), so I don't know how to explain the variance...