@palimondo 2018-12-03
@lorentey, you didn't answer my question from [older conversation](https://github.com/apple/swift/pull/20769#discussion_r237246638):
> longASCII and shortASCII are quite different workloads though.

How so? I only see the same string repeated different times. That seems to me just like a different way to scale the workload size, instead of setting the inner-loop multiplier. Am I mistaken? Is the difference about some kind of small string optimization behind the scenes? If so, this should probably be also documented in the comments.