@palimondo 2019-02-14
> 🧐`longMixed` look great, but `longASCII` is weirding me out… You said it looked fine locally, so I don't know how to explain the variance...

Upon further thought, the variance can be explained this way: `run_smoke_bench` initially gathers only 3 samples per benchmark and that's all we get for the added benchmarks, therefore any benchmarks that have _any_ warmup will manifest itself as this variance. 
@eeckstein: _We should gather more independent samples also for the added benchmarks_.

----

@lorentey @milseman There's something strange going on with these benchmarks and/or the breadcrumbs implementation. 

After Karoy ran the `Benchmark_Driver check` locally he sized workloads and iterations of the benchmarks to sit nicely in the hundreds of μs range, see [`StringBreadcrumbs` benchmark report](https://github.com/apple/swift/pull/20769#issuecomment-441849354) above. Then there were two force-pushes that are lost to history, which probably re-adjusted the workloads of `CopyUTF16CodeUnits`, because in the last benchmark [report above](https://github.com/apple/swift/pull/20769#issuecomment-442596088), their runtime dropped by a factor of 10. (We should IMO revert this by increasing the `count` value back to `5_000`.)

TEST | MIN | MAX | MEAN |
--|--|--|--
Breadcrumbs.CopyUTF16CodeUnits.ASCII | 47 | 48 | 48
Breadcrumbs.CopyUTF16CodeUnits.Mixed | 52 | 55 | 53

Strangely, [as soon as 3rd of December](https://github.com/apple/swift/pull/20965#issuecomment-443785665) some of these start to sporadically appear in the `-O` Performance  changes reported by @swift-ci start with very low runtimes:

TEST | OLD | NEW | DELTA| RATIO
--|--|--|--|--
Breadcrumbs.CopyUTF16CodeUnits.ASCII | 20 | 18 | -10.0% | 1.11x
Breadcrumbs.CopyUTF16CodeUnits.Mixed | 68 | 55 | -19.1% | 1.24x

There were no benchmark reports from @swift-ci in any PR's here on GitHub that would include the improvement as intentional change…  I have noticed that something is strange here during my janitor duty, as even on my measly machine the `MutatedIdxToUTF16.ASCII ` and `MutatedUTF16ToIdx.ASCII` dropped to 15 μs and were picked up by the `BenchmarkDoctor` heuristic that flags runtimes below 20 μs (because than even 1 μs change is above the 5% threshold increasing the chances it shows up in report).

I've looked at it in Instruments: 

ASCII workload's stack trace:
````
   2 Benchmark_O 289.0  MutatedIdxToUTF16.run(iterations:)
   1 libswiftCore.dylib 35.0  StringProtocol._toUTF16Offset(_:)
   0 libswiftCore.dylib 12.0  specialized String.UTF16View._nativeGetOffset(for:)
````

Mixed workload's stack trace:
````
  10 Benchmark_O 16815.0  MutatedIdxToUTF16.run(iterations:)
   9 libswiftCore.dylib 15682.0  StringProtocol._toUTF16Offset(_:)
   8 libswiftCore.dylib 15660.0  specialized String.UTF16View._nativeGetOffset(for:)
   7 libswiftCore.dylib 15342.0  specialized _StringGuts.populateBreadcrumbs(_:)
   6 libswiftCore.dylib 15012.0  specialized _StringBreadcrumbs.init(_:)
````

It looks like ASCII workload is not using any breadcrumbs. Is that normal or a bug?!